RESTRUCTURED: source code
This commit goes with source code restructured.
