+++ Sources/PerseusGeoLocationKit/Models/Commons.swift
@@ -1,0 +1,65 @@
+//
+//  Commons.swift
+//  PerseusGeoLocationKit
+//
+//  Created by Mikhail Zhigulin in 7531.
+//
+//  Copyright © 7531 Mikhail Zhigulin of Novosibirsk.
+//
+//  Licensed under the MIT license. See LICENSE file.
+//  All rights reserved.
+//
+
+import Foundation
+import CoreLocation
+
+// MARK: - Constants
+
+public let APPROPRIATE_ACCURACY = LocationAccuracy.threeKilometers
+
+// MARK: - Result
+
+public enum Result<Value, Error: Swift.Error> {
+
+    case success(Value)
+    case failure(Error)
+}
+
+// MARK: - Notifications
+
+extension Notification.Name {
+
+    // Current Location
+    public static let locationDealerCurrentNotification =
+    Notification.Name("locationDealerCurrentNotification")
+
+    // Location Changing Updates
+    public static let locationDealerUpdatesNotification =
+    Notification.Name("locationDealerUpdatesNotification")
+
+    // Error
+    public static let locationDealerErrorNotification =
+    Notification.Name("locationDealerErrorNotification")
+
+    // Location Service Status
+    public static let locationDealerStatusChangedNotification =
+    Notification.Name("locationDealerStatusChangedNotification")
+}
+
+extension CLAuthorizationStatus: CustomStringConvertible {
+
+    public var description: String {
+        switch self {
+        case .notDetermined:
+            return "notDetermined"
+        case .restricted:
+            return "restricted"
+        case .denied:
+            return "denied"
+        case .authorizedAlways:
+            return "authorizedAlways"
+        case .authorizedWhenInUse: // iOS only.
+            return "authorizedWhenInUse"
+        }
+    }
+}
