+++ Sources/PerseusGeoLocationKit/Models/LocationDealerPermit.swift
@@ -71,21 +71,3 @@ public func getPermit(serviceEnabled: Bool,
 
     return .allowed
 }
-
-extension CLAuthorizationStatus: CustomStringConvertible {
-
-    public var description: String {
-        switch self {
-        case .notDetermined:
-            return "notDetermined"
-        case .restricted:
-            return "restricted"
-        case .denied:
-            return "denied"
-        case .authorizedAlways:
-            return "authorizedAlways"
-        case .authorizedWhenInUse: // iOS only.
-            return "authorizedWhenInUse"
-        }
-    }
-}
